Fix error on Newest Nuxt Version
- Import nuxt module inside component for navigation-list.js & Make the list more accessable
- Rename and move /config dir to /utils to satisfy auto import on new nuxt version
