+++ package.json
@@ -14,7 +14,7 @@
     "@nuxt/image-edge": "^1.0.0-27968280.9739e4d",
     "@nuxtjs/i18n": "^8.0.0-beta.11",
     "@unlighthouse/nuxt": "^0.8.6",
-    "nuxt": "^3.2.2",
+    "nuxt": "^3.6.1",
     "nuxt-gtag": "^0.5.7",
     "nuxt-seo-kit": "^1.3.8",
     "typescript": "^5.1.3"
