+++ storyblok/SectionFooter.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { listCameras, listProductivity, listFeatures, listAbout, listPartners, listApp, listSocials } from "../config/navigation-list";
+import { navigationList } from "../utils/navigation-list";
 const props = defineProps({
   blok: Object,
   phone_number: String,
@@ -389,7 +389,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
           <h3>Construction Cameras</h3>
           <ul class="">
             <li
-              v-for="(item, i) in listCameras"
+              v-for="(item, i) in navigationList().listCameras"
               :key="i"
               class="dropdown"
               :id="item.id"
@@ -400,7 +400,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
           <h3>Productivity</h3>
           <ul class="">
             <li
-              v-for="(item, i) in listProductivity"
+              v-for="(item, i) in navigationList().listProductivity"
               :key="i"
               class="dropdown"
               :id="item.id"
@@ -413,7 +413,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
           <h3>About</h3>
           <ul class="">
             <li
-              v-for="(item, i) in listAbout"
+              v-for="(item, i) in navigationList().listAbout"
               :key="i"
               class="dropdown"
               :id="item.id"
@@ -424,7 +424,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
           <h3>Partners</h3>
           <ul class="">
             <li
-              v-for="(item, i) in listPartners"
+              v-for="(item, i) in navigationList().listPartners"
               :key="i"
               class="dropdown"
               :id="item.id"
@@ -437,7 +437,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
           <h3>Features</h3>
           <ul class="">
             <li
-              v-for="(item, i) in listFeatures"
+              v-for="(item, i) in navigationList().listFeatures"
               :key="i"
               class="dropdown"
               :id="item.id"
@@ -453,7 +453,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
         <div class="row">
           <div class="col-md-3 align-self-center slideLeft">
             <a
-              v-for="(item, i) in listApp"
+              v-for="(item, i) in navigationList().listApp"
               :key="i"
               :href="item.href || 'javascript:void(0)'"
               target="_blank"
@@ -803,7 +803,7 @@ const telLink = computed(() => `tel:+${phoneNumber}`);
             class="col-md-4 align-self-center slideRight mobile-social-icons"
           >
             <ul class="social-links">
-              <li v-for="(item, i) in listSocials" :key="i">
+              <li v-for="(item, i) in navigationList().listSocials" :key="i">
                 <a
                   :href="item.href || 'javascript:void(0)'"
                   target="_blank"
